+++ .eslintrc
@@ -12,7 +12,7 @@
 		"node": true
 	},
 	"globals": {
-		"Config": false, "Monitor": false, "toId": false, "Tools": false, "LoginServer": false,
+		"Config": false, "Monitor": false, "toId": false, "Dex": false, "LoginServer": false,
 		"Users": false, "Punishments": false, "Rooms": false, "Verifier": false, "Chat": false,
 		"Tournaments": false, "Dnsbl": false, "Sockets": false, "TeamValidator": false, "Ladders": false
 	},
