+++ mods/gen1/scripts.js
@@ -32,7 +32,7 @@ exports.BattleScripts = {
 			return this.modifiedStats[statName];
 		},
 		// Gen 1 function to apply a stat modification that is only active until the stat is recalculated or mon switched.
-		// Modified stats are declared in BattlePokemon object in battle-engine.js in about line 303.
+		// Modified stats are declared in the Pokemon object in sim/pokemon.js in about line 681.
 		modifyStat: function (stat, modifier) {
 			if (!(stat in this.stats)) return;
 			this.modifiedStats[stat] = this.battle.clampIntRange(Math.floor(this.modifiedStats[stat] * modifier), 1, 999);
@@ -582,7 +582,7 @@ exports.BattleScripts = {
 
 		return damage;
 	},
-	// boost can be found on battle-engine.js on Battle object.
+	// boost can be found on sim/battle.js on Battle object.
 	// It deals with Pokémon stat boosting, including Gen 1 buggy behaviour with burn and paralyse.
 	boost: function (boost, target, source, effect) {
 		if (this.event) {
@@ -627,7 +627,7 @@ exports.BattleScripts = {
 		}
 		this.runEvent('AfterBoost', target, source, effect, boost);
 	},
-	// damage can be found in battle-engine.js on the Battle object. Not to confuse with BattlePokemon.prototype.damage
+	// damage can be found in sim/battle.js on the Battle object. Not to confuse with BattlePokemon.prototype.damage
 	// It calculates and executes the damage damage from source to target with effect.
 	// It also deals with recoil and drains.
 	damage: function (damage, target, source, effect) {
@@ -689,7 +689,7 @@ exports.BattleScripts = {
 
 		return damage;
 	},
-	// directDamage can be found on battle-engine.js in Battle object
+	// directDamage can be found on sim/battle.js in Battle object
 	// It deals direct damage damage from source to target with effect.
 	// It also deals with Gen 1 weird Substitute behaviour.
 	directDamage: function (damage, target, source, effect) {
@@ -730,7 +730,7 @@ exports.BattleScripts = {
 
 		return damage;
 	},
-	// getDamage can be found on battle-engine.js on the Battle object.
+	// getDamage can be found on sim/battle.js on the Battle object.
 	// It calculates the damage pokemon does to target with move.
 	getDamage: function (pokemon, target, move, suppressMessages) {
 		// First of all, we get the move.
@@ -1126,7 +1126,7 @@ exports.BattleScripts = {
 			// Spammable attacks are: Thunderbolt, Psychic, Surf, Blizzard, Earthquake.
 			let pokemonWeaknesses = [];
 			for (let type in weaknessCount) {
-				let increaseCount = Tools.getImmunity(type, template) && Tools.getEffectiveness(type, template) > 0;
+				let increaseCount = Dex.getImmunity(type, template) && Dex.getEffectiveness(type, template) > 0;
 				if (!increaseCount) continue;
 				if (weaknessCount[type] >= 2) {
 					skip = true;
