+++ sim/prng.js
