+++ test/simulator/abilities/desolateland.js
@@ -25,7 +25,7 @@ describe('Desolate Land', function () {
 		const attacker = battle.p1.active[0];
 		const defender = battle.p2.active[0];
 		assert.hurtsBy(defender, 152, () => battle.commitDecisions());
-		const move = Tools.getMove('incinerate');
+		const move = Dex.getMove('incinerate');
 		const basePower = battle.runEvent('BasePower', attacker, defender, move, move.basePower, true);
 		assert.strictEqual(basePower, move.basePower);
 	});
