+++ test/simulator/misc/weather.js
@@ -18,7 +18,7 @@ describe('Weather damage calculation', function () {
 		const attacker = battle.p1.active[0];
 		const defender = battle.p2.active[0];
 		assert.hurtsBy(defender, 152, () => battle.commitDecisions());
-		const move = Tools.getMove('incinerate');
+		const move = Dex.getMove('incinerate');
 		const basePower = battle.runEvent('BasePower', attacker, defender, move, move.basePower, true);
 		assert.strictEqual(basePower, move.basePower);
 	});
@@ -31,7 +31,7 @@ describe('Weather damage calculation', function () {
 		const attacker = battle.p1.active[0];
 		const defender = battle.p2.active[0];
 		assert.hurtsBy(defender, 50, () => battle.commitDecisions());
-		const move = Tools.getMove('incinerate');
+		const move = Dex.getMove('incinerate');
 		const basePower = battle.runEvent('BasePower', attacker, defender, move, move.basePower, true);
 		assert.strictEqual(basePower, move.basePower);
 	});
