+++ test/simulator/moves/ragepowder.js
@@ -2,7 +2,7 @@
 
 const assert = require('./../../assert');
 const common = require('./../../common');
-const BattleEngine = require('./../../../battle-engine');
+const Sim = require('./../../../sim');
 
 let battle;
 
@@ -28,15 +28,15 @@ describe('Rage Powder', function () {
 		let hitCount = [0, 0, 0];
 		battle.p1.active[0].damage = function (...args) {
 			hitCount[0]++;
-			return BattleEngine.BattlePokemon.prototype.damage.apply(this, args);
+			return Sim.Pokemon.prototype.damage.apply(this, args);
 		};
 		battle.p1.active[1].damage = function (...args) {
 			hitCount[1]++;
-			return BattleEngine.BattlePokemon.prototype.damage.apply(this, args);
+			return Sim.Pokemon.prototype.damage.apply(this, args);
 		};
 		battle.p1.active[2].damage = function (...args) {
 			hitCount[2]++;
-			return BattleEngine.BattlePokemon.prototype.damage.apply(this, args);
+			return Sim.Pokemon.prototype.damage.apply(this, args);
 		};
 		battle.choose('p2', 'move 1 2, move 1 2, move 1 2');
 		battle.commitDecisions();
@@ -60,15 +60,15 @@ describe('Rage Powder', function () {
 		let hitCount = [0, 0, 0];
 		battle.p1.active[0].damage = function (...args) {
 			hitCount[0]++;
-			return BattleEngine.BattlePokemon.prototype.damage.apply(this, args);
+			return Sim.Pokemon.prototype.damage.apply(this, args);
 		};
 		battle.p1.active[1].damage = function (...args) {
 			hitCount[1]++;
-			return BattleEngine.BattlePokemon.prototype.damage.apply(this, args);
+			return Sim.Pokemon.prototype.damage.apply(this, args);
 		};
 		battle.p1.active[2].damage = function (...args) {
 			hitCount[2]++;
-			return BattleEngine.BattlePokemon.prototype.damage.apply(this, args);
+			return Sim.Pokemon.prototype.damage.apply(this, args);
 		};
 		battle.choose('p2', 'move 1 3, move 1 1, move 1 1');
 		battle.commitDecisions();
