+++ test/simulator/moves/roost.js
@@ -2,7 +2,7 @@
 
 const assert = require('./../../assert');
 const common = require('./../../common');
-const BattleEngine = require('./../../../battle-engine');
+const Sim = require('./../../../sim');
 
 let battle;
 
@@ -67,7 +67,7 @@ describe('Roost', function () {
 		let hitCount = 0;
 		battle.p2.active[0].damage = function (...args) {
 			hitCount++;
-			return BattleEngine.BattlePokemon.prototype.damage.apply(this, args);
+			return Sim.Pokemon.prototype.damage.apply(this, args);
 		};
 
 		battle.commitDecisions();
